Add command for changing the master key
It was very inconvenient and error-prone to migrate `onetimepass` to
another device without having such command: it required to manually copy
the database, and manually insert the master key into the another
device's keychain (!).

Implementation introduced in this commit is rather crude, but works good
enough as an initial version.